Updated Embree (OSX/Win) to 2.7, TBB (OSX/Win) to 4.4

+++ include/embree2/rtcore_ray.h
@@ -1,5 +1,5 @@
 // ======================================================================== //
-// Copyright 2009-2014 Intel Corporation                                    //
+// Copyright 2009-2015 Intel Corporation                                    //
 //                                                                          //
 // Licensed under the Apache License, Version 2.0 (the "License");          //
 // you may not use this file except in compliance with the License.         //
@@ -26,10 +26,10 @@ struct RTCORE_ALIGN(16)  RTCRay
   /* ray data */
 public:
   float org[3];      //!< Ray origin
-  int   align0;
+  float align0;
   
   float dir[3];      //!< Ray direction
-  int   align1;
+  float align1;
   
   float tnear;       //!< Start of ray segment
   float tfar;        //!< End of ray segment (set to hit distance)
@@ -40,7 +40,7 @@ public:
   /* hit data */
 public:
   float Ng[3];       //!< Unnormalized geometry normal
-  int   align2;
+  float align2;
   
   float u;           //!< Barycentric u coordinate of hit
   float v;           //!< Barycentric v coordinate of hit
